Update 9
Fixed the restart button so it restarts the game, you can now get 10 points if you land safely and if you land safely and on top of a person you can get 50 points, if you don't land safely, you get 0 points. The score is then written in the lose and win screen.